Show attached file size, show file status (whether it exceeds board limitations). Fix a bug in 2ch.hk passcode info request handling where we would store returned "max_file_size" value in KB instead of converting it into bytes first.

+++ Kuroba/app/src/main/java/com/github/k1rakishou/chan/core/site/limitations/PasscodePostingLimitationsInfo.kt
@@ -2,5 +2,5 @@ package com.github.k1rakishou.chan.core.site.limitations
 
 class PasscodePostingLimitationsInfo(
   val maxAttachedFilesPerPost: Int,
-  val maxTotalAttachablesSize: Int
+  val maxTotalAttachablesSize: Long
 )
